+++ Final Assignment/Final Assignment_DOUBLE/A2Matrix_DOUBLE/add_sub_product_DOUBLE.c
@@ -1,4 +1,6 @@
 #include "A2Matrix_DOUBLE.h"
+
+//Calculate Matrix1 + Matrix2
 int addition_DOUBLE(double matrix1[30][30], double matrix2[30][30], int R1, int C1, int R2, int C2, double matrix3[30][30])
 {
     int i,j;
@@ -16,6 +18,7 @@ int addition_DOUBLE(double matrix1[30][30], double matrix2[30][30], int R1, int 
         return 0;}
 }
 
+//Calculate Matrix1  Matrix2
 int subtraction_DOUBLE(double matrix1[30][30], double matrix2[30][30], int R1, int C1, int R2, int C2, double matrix3[30][30])
 {
     int i,j;
