+++ Final Assignment/Final Assignment_DOUBLE/A2Matrix_DOUBLE/delRCcoFactor_DOUBLE.c
@@ -13,3 +13,33 @@ int delRCcoFactor_DOUBLE(double matrix[30][30], int R, int C, int dR, int dC, do
     }
 
 }
+
+//delete row of a matrix to make new one
+void delRow_DOUBLE(double matrix[30][30],int R, int C, int dR, double temp[30][30])
+{
+    int i,j;
+
+    for(i=1;i<dR;i++)
+        for(j=1;j<=C;j++)
+            temp[i][j]=matrix[i][j];
+    for(i=dR;i<=R;i++)
+        for(j=1;j<=C;j++)
+            temp[i][j]=matrix[i+1][j];
+
+}
+
+//delete column of a matrix to make new one
+void delCol_DOUBLE(double matrix[30][30],int R, int C, int dC, double temp[30][30])
+{
+    int i,j;
+
+    for(i=1;i<=R;i++)
+        for(j=1;j<dC;j++)
+            temp[i][j]=matrix[i][j];
+
+    for(i=1;i<=R;i++)
+        for(j=dC;j<=C;j++)
+            temp[i][j]=matrix[i][j+1];
+
+}
+
