+++ Final Assignment/Final Assignment_DOUBLE/A2Matrix_DOUBLE/det_DOUBLE.c
@@ -1,6 +1,8 @@
 //determinant of m x m matrix
 #include <math.h>
 #include "A2Matrix_DOUBLE.h"
+
+//Calculate the determinant of matrix
 double det_DOUBLE (double matrix[30][30],int R, int C )
 {
     int delRCcoFactor_DOUBLE(double matrix[30][30], int R, int C, int dR, int dC, double temp[30][30]);
@@ -27,6 +29,24 @@ double det_DOUBLE (double matrix[30][30],int R, int C )
    }
     //printf("%d\t", det);
    return det;
+}
+
+//determinant of a 2x2 matrix
+
+double det2_DOUBLE(double matrix[30][30], int R, int C)
+{
+    if (R!=C||R!=2||C!=2)
+        {
+            printf("\n\n det2_DOUBLE gets some higher order of matrix\n\n");
+            return 0;
+
+        }
+    else
+    {
+        double det2=0;
+        det2=matrix[1][1]*matrix[2][2]-matrix[2][1]*matrix[1][2];
+        return det2;
+    }
 }
