--- Final Assignment/Final Assignment_DOUBLE/A2Matrix_DOUBLE/enterRC_DOUBLE.c
@@ -1,13 +1,0 @@
-#include "A2Matrix_DOUBLE.h"
-void enterRC_DOUBLE(int matrix[30][30],int *R, int *C)
-{
-    int r,c;
-    printf("Enter row of %s: ",matrix[0][0]);
-
-    printf("Enter column of %s: ",matrix[0][0]);
-    scanf("%d",&c);
-
-    *R=r;
-    *C=c;
-
-}
